pokemon pictures showing in the UI
