+++ js/data/pokedex.js
